Converted Commands to Menu Actions
Converted FireTeam to MonoBehaviour and made it follow the center of mass of the formation.
Added FireTeam prefab that is clickable and draws a projection texture to denote collider bounds.
Changed AllyGameController to use FireTeam prefab.
Made MenuAction abstract and added 3 abstract functions: GetSelectedType, GetTargetType and Execute.
Added Projector Standard Assets to the project.
Added MoveTeamMenuAction and ChangeFormationMenuAction classes, which are conversions of their respective commands.
Added 4 actions to MenuPrefab: Wedge, File, Cover and Move.