Converted Attack and Disengage Commands to Menu Actions
Added a GetAllMembers() function to FireTeam class that returns all members, including the leader.
Changed all previous MenuActions to reference Command code.
Added TeamAttackEnemyCommand and TeamDisengageCommand that reuses Attack and Disengage commands and relays them to the team.
Added TeamAttackEnemyMenuAction and TeamDisengageMenuAction that wraps the beforementioned classes and uses them for menu.
Added the new Menu Actions to the menu prefab.