Added very rudimentary map mode; characters can have accurate positions reflected in map mode; just need a means to access the list of all units on the scene.
To select a unit from this map mode; we just need to combine the top down bird's eye view controller with this; essentially have a camera with a orthographic camera position in a central location but with its camera component disabled and is pointed down.

Disable all camera's, enable birds eye camera.
When on click in the map mode, draw a ray to see if a unit is selected.
Then reenable camera/disable bird's eye camera.

Thus, we have the best of all world's.

Right clicking to send orders is done by a similar logic; since this is all done within the span of a single frame; and since the user is in a map mode anyways; there's zero notice by the player that the camera's were just juggled.